# Conflicts:
#	clients_app/src/Controller/AppController.php
#	clients_app/src/Entity/Bilety.php
#	clients_app/src/Entity/Uzytkownicy.php
#	workers_app/src/Entity/Bilety.php
#	workers_app/src/Entity/Uzytkownicy.php

+++ workers_app/src/Controller/PromotionsController.php
@@ -46,6 +46,26 @@ class PromotionsController extends Controller
     }
 
     /**
+     * @Route("/promotions/old/{page<[1-9]\d*>?1}", name="workers_app/promotions/old", methods={"GET"})
+     */
+    public function old($page)
+    {
+        if($this->isGranted('IS_AUTHENTICATED_FULLY')) {
+            $pageLimit = $this->getParameter('page_limit');
+            $pageCount = $this->getDoctrine()->getRepository(Promocje::class)->getPageCountOfOld($pageLimit);
+
+            if($page > $pageCount and $pageCount != 0)
+                return $this->redirectToRoute('workers_app/promotions/old');
+            else {
+                $promotions = $this->getDoctrine()->getRepository(Promocje::class)->findOld($page, $pageLimit);
+                return $this->render('workersApp/promotions/listOld.html.twig', array('promotions' => $promotions, 'currentPage' => $page, 'pageCount' => $pageCount));
+            }
+        } else {
+            return $this->redirectToRoute('workers_app/login_page');
+        }
+    }
+
+    /**
      * @Route("/promotions/add", name="workers_app/promotions/add", methods={"GET", "POST"})
      */
     public function add(Request $request)
@@ -89,7 +109,7 @@ class PromotionsController extends Controller
         if($this->isGranted('ROLE_MANAGER') or $this->isGranted('ROLE_ADMIN')) {
             $promotion = $this->getDoctrine()->getRepository(Promocje::class)->find($id);
 
-            if($promotion->getPoczatekpromocji()->format("Y-m-d") <= date("Y-m-d"))
+            if($promotion == null or $promotion->getPoczatekpromocji()->format("Y-m-d") <= date("Y-m-d"))
                 throw new NotFoundHttpException();
 
             $form = $this->getForm($promotion);
@@ -209,7 +229,7 @@ class PromotionsController extends Controller
         }
         if($this->isGranted('ROLE_MANAGER') or $this->isGranted('ROLE_ADMIN')) {
             $promotion = $this->getDoctrine()->getRepository(Promocje::class)->find($id);
-            if($promotion->getPoczatekpromocji()->format("Y-m-d") > date("Y-m-d")) {
+            if($promotion != null and $promotion->getPoczatekpromocji()->format("Y-m-d") > date("Y-m-d")) {
                 $entityManager = $this->getDoctrine()->getManager();
 
                 $entityManager->remove($promotion);
